requirements.txt
alembic==1.13.1
fastapi==0.111.0
httpx==0.27.0
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-jose==3.3.0
SQLAlchemy==2.0.30
uvicorn==0.30.1
gunicorn==22.0.0
cryptography==42.0.5
